refactor: extract Path, Anchor functions into hdi + hdk
