+++ crates/holochain_integrity_types/Cargo.toml
@@ -36,6 +36,7 @@ tracing = { version = "0.1", optional = true }
 [dev-dependencies]
 holochain_integrity_types = { path = ".", features = ["test_utils"]}
 serde_json = "1"
+fixt = { path = "../fixt" }
 
 [features]
 default = []
@@ -43,6 +44,8 @@ default = []
 full = ["default", "subtle-encoding", "kitsune_p2p_timestamp/full"]
 
 full-dna-def = ["derive_builder"]
+
+fixturators = ["holo_hash/fixturators"]
 
 test_utils = [
   "full",
@@ -53,4 +56,5 @@ test_utils = [
   "holo_hash/hashing",
   "holo_hash/test_utils",
   "holochain_serialized_bytes/arbitrary",
+  "fixturators",
 ]
