+++ crates/holochain_integrity_types/src/hash_path/shard.rs
@@ -33,6 +33,7 @@ impl ShardStrategy {
     }
 }
 
+/// Errors for parsing a string into a sharded path
 #[derive(Debug)]
 pub enum ParseShardStrategyError {
     /// Could not parse the shard depth.
